Moving results of typographic first discussion to dev branch
Given the technical implications the results are yet missing on the
main use case due to limitations of web fonts.
As this is a public guideline, I prefer to hold back on this a
little further until we've got agreement how to address those issues
in accordance with our recommendations.