Added a target property
I discovered when the "Contribute on GitHub button is clicked, it just loads closing up the file the individual was actually reading which is not suppose to be. I added a target property so it opens up a new tab whenever it is clicked, in that case the user can still continue what he/she was doing previously before clicking the button so he/she does not loose track of where he/she was.