Added a target property on GitHub link
I discovered when the "Contribute on GitHub" button is clicked, it just loads closing up the file the individual was actually reading which is not suppose to be. I added a target property so it opens up a new tab whenever it is clicked, in that case the user can still continue what they were doing previously before clicking the button so they do not loose track of where they were.